testmongo/requirements.txt
asgiref==3.5.2
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.3
django-cors-headers==3.13.0
django-import-export==3.0.1
django-rest-swagger==2.2.0
django-restframework-apiview==1.4.3
django-storages==1.13.2
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
drf-yasg==1.21.4
et-xmlfile==1.1.0
google-api-core==2.11.0
google-api-python-client==2.70.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
gunicorn==20.1.0
httplib2==0.21.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jwt==1.3.1
MarkupPy==1.14
MarkupSafe==2.1.1
numpy==1.24.2
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.10
packaging==21.3
pandas==1.5.3
Pillow==9.3.0
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pymongo==3.12.3
pyparsing==3.0.9
pyshorteners==1.0.1
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
qrcode==7.3.1
requests==2.28.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.17.6
six==1.16.0
sqlparse==0.2.4
tablib==3.2.1
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.12
xlrd==2.0.1
xlwt==1.3.0
